Great re-working of components
