+++ .gitignore
@@ -15,6 +15,7 @@
 .env.development.local
 .env.test.local
 .env.production.local
+/.idea
 
 npm-debug.log*
 yarn-debug.log*
