+++ src/Players/player.css
@@ -1,0 +1,108 @@
+.player-list {
+    margin-top: 60px;
+    padding-top: 20px;
+}
+
+.player-status-bar {
+    position: fixed;
+    box-shadow: 0 5px rgba(0, 0, 0, 0.1);
+    width: 100%;
+    top:0;
+    margin-bottom: 20px;
+    background: cornflowerblue;
+    height: 60px;
+    color: whitesmoke;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    transition: .5s linear all;
+    z-index: 1;
+}
+.player-status-bar .numbers {
+    font-size: 2rem;
+    font-family: 'PT Sans', sans-serif;
+    margin-right: 10px;
+}
+.player-status-bar .playing {
+    transition: .5s linear all;
+}
+.player-status-bar.highlight .playing {
+    font-size: 3rem;
+}
+
+.player-group-name {
+    text-transform: Uppercase;
+    letter-spacing: 6vw;
+    text-indent: 6vw;
+    font-size: 3vw;
+    color: #ed9d64;
+    overflow: hidden;
+}
+@media screen and (min-width: 650px) {
+    .player-group-name {
+        font-size: 1.1rem;
+        text-indent: 3rem;
+        letter-spacing: 3rem;
+    }
+}
+
+.player-group {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    flex-wrap: wrap;
+}
+
+.player {
+    width: 120px;
+    margin: 10px;
+    flex: 1 0 calc(15% - 10px);
+    transform: scale(0.9);
+    transition: all 0.3s;
+    color: #333;
+}
+
+.player:hover {
+    transform: scale(1);
+}
+
+.player-photo {
+    margin: 0 auto;
+    width: 120px;
+    height: 120px;
+    background-image: url('avatar.svg');
+    background-size: 120px 120px;
+    background-repeat: no-repeat;
+    filter: grayscale(100%);
+    border: 7px solid rgba(100,237,189,.4);
+    border-radius:50%;
+    cursor: pointer;
+    overflow: hidden;
+}
+.player-photo img {
+    max-width: 120px;
+}
+
+.player h3, .player input {
+    margin: 5px 0 0;
+    color: cornflowerblue;
+    font-size: 1.17em;
+    font-weight: bold;
+    transition: all 0.3s;
+}
+.player input:focus {
+    border-bottom: 1px solid cornflowerblue;
+    outline: none;
+    margin: 10px 0;
+}
+
+.player input {
+    background: transparent;
+    border: none;
+    text-align: center;
+    width:100%;
+}
+
+.player.playing .player-photo {
+    filter: unset;
+}
